Reduced base classes down even more.
