+++ CS6502.Core/Instructions/Memory/CMP.cs
