+++ CS6502.Core/Instructions/Memory/CPY.cs
